Add test rule, add out folder to gitignore

+++ java_skeleton_generation/src/util/Position.java
@@ -1,25 +1,25 @@
 package util;
 
 public class Position {
-    private float x;
-    private float y;
-    private float z;
+    private double x;
+    private double y;
+    private double z;
 
-    public Position(float x, float y, float z) {
+    public Position(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public float getZ() {
+    public double getZ() {
         return z;
     }
 }
